# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tu vecino te ayuda API",
        "description": "This is a Beta version of tuvecinoteayuda api. Use it carefully and report all bugs and security problems found.",
        "termsOfService": "https://tuvecinoteayuda.org/politica_privacidad.pdf",
        "contact": {
            "email": "hablamos@tuvecinoteayuda.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://www.api.tuvecinoteayuda.org/",
            "description": "Tu vecino te Ayuda OpenApi host"
        }
    ],
    "paths": {
...
        "/api/public/auth/password/email": {
            "post": {
                "tags": [
                    "password",
                    "user"
                ],
                "summary": "Request reset email",
                "operationId": "sendResetLinkEmail",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email, required",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Send a reset link to the given user.",
                                            "type": "string"
                                        },
                                        "status_code": {
                                            "description": "Send a reset link to the given user.",
                                            "type": "integer"
                                        },
                                        "status": {
                                            "description": "Send a reset link to the given user.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error ocurred"
                    },
                    "422": {
                        "description": "Request invalid. see errors"
                    }
                }
            }
        },
...
    },
...
}